requirements.txt
Django==1.9.6
djangorestframework==3.5.1
django-cors-headers (1.2.2)